+++ r-diagnostics.tf
@@ -9,18 +9,8 @@ resource "azurerm_virtual_machine_scale_set_extension" "azure_monitor_agent" {
   virtual_machine_scale_set_id = azurerm_linux_virtual_machine_scale_set.linux_vmss.id
 }
 
-resource "null_resource" "azure_monitor_link" {
-  provisioner "local-exec" {
-    command = <<EOC
-      az rest --subscription ${data.azurerm_client_config.current.subscription_id} \
-              --method PUT \
-              --url https://management.azure.com${azurerm_linux_virtual_machine_scale_set.linux_vmss.id}/providers/Microsoft.Insights/dataCollectionRuleAssociations/${azurerm_linux_virtual_machine_scale_set.linux_vmss.name}-dcrassociation?api-version=2019-11-01-preview \
-              --body '{"properties":{"dataCollectionRuleId": "${var.azure_monitor_data_collection_rule_id}"}}'
-EOC
-  }
-
-  triggers = {
-    dcr_id  = var.azure_monitor_data_collection_rule_id
-    vmss_id = azurerm_linux_virtual_machine_scale_set.linux_vmss.id
-  }
+resource "azurerm_data_collection_rule_association" "azure_monitor" {
+  name                    = "dcra-${azurerm_linux_virtual_machine_scale_set.linux_vmss.name}"
+  target_resource_id      = azurerm_linux_virtual_machine_scale_set.linux_vmss.id
+  data_collection_rule_id = var.azure_monitor_data_collection_rule_id
 }